+++ tests/test_gui.py
@@ -62,7 +62,7 @@ def locale_error_zenity(gui_provider, monkeypatch):
 
 
 class TestSelectApp:
-    def test_select_game(self, gui_provider, steam_app_factory):
+    def test_select_game(self, gui_provider, steam_app_factory, steam_dir):
         """
         Select a game using the GUI
         """
@@ -73,15 +73,45 @@ class TestSelectApp:
 
         # Fake user selecting 'Fake game 2'
         gui_provider.mock_stdout = "Fake game 2: 20"
-        steam_app = select_steam_app_with_gui(steam_apps=steam_apps)
+        steam_app = select_steam_app_with_gui(
+            steam_apps=steam_apps, steam_path=steam_dir
+        )
 
         assert steam_app == steam_apps[1]
 
-        # Check that choices were displayed
-        assert b"Fake game 1: 10\nFake game 2: 20" \
-            in gui_provider.kwargs["input"]
+        input_ = gui_provider.kwargs["input"]
 
-    def test_select_game_no_choice(self, gui_provider, steam_app_factory):
+        # Check that choices were displayed
+        assert b"Fake game 1: 10\n" in input_
+        assert b"Fake game 2: 20" in input_
+
+    def test_select_game_icons(
+            self, gui_provider, steam_app_factory, steam_dir):
+        """
+        Select a game using the GUI. Ensure that icons are used in the dialog
+        whenever available.
+        """
+        steam_apps = [
+            steam_app_factory(name="Fake game 1", appid=10),
+            steam_app_factory(name="Fake game 2", appid=20),
+            steam_app_factory(name="Fake game 3", appid=30),
+        ]
+
+        # Create icons for game 1 and 3
+        (steam_dir / "appcache" / "librarycache" / "10_icon.jpg").touch()
+        (steam_dir / "appcache" / "librarycache" / "30_icon.jpg").touch()
+
+        gui_provider.mock_stdout = "Fake game 2: 20"
+        select_steam_app_with_gui(steam_apps=steam_apps, steam_path=steam_dir)
+
+        input_ = gui_provider.kwargs["input"]
+
+        assert b"librarycache/10_icon.jpg\nFake game 1" in input_
+        assert b"icon_placeholder.png\nFake game 2" in input_
+        assert b"librarycache/30_icon.jpg\nFake game 3" in input_
+
+    def test_select_game_no_choice(
+            self, gui_provider, steam_app_factory, steam_dir):
         """
         Try choosing a game but make no choice
         """
@@ -91,12 +121,14 @@ class TestSelectApp:
         gui_provider.mock_stdout = ""
 
         with pytest.raises(SystemExit) as exc:
-            select_steam_app_with_gui(steam_apps=steam_apps)
+            select_steam_app_with_gui(
+                steam_apps=steam_apps, steam_path=steam_dir
+            )
 
         assert exc.value.code == 0
 
     def test_select_game_broken_zenity(
-            self, broken_zenity, monkeypatch, steam_app_factory):
+            self, broken_zenity, monkeypatch, steam_app_factory, steam_dir):
         """
         Try choosing a game with a broken Zenity executable that
         prints a specific error message that Protontricks knows how to ignore
@@ -110,12 +142,13 @@ class TestSelectApp:
 
         # Fake user selecting 'Fake game 2'
         broken_zenity.mock_stdout = "Fake game 2: 20"
-        steam_app = select_steam_app_with_gui(steam_apps=steam_apps)
+        steam_app = select_steam_app_with_gui(
+            steam_apps=steam_apps, steam_path=steam_dir)
 
         assert steam_app == steam_apps[1]
 
     def test_select_game_locale_error(
-            self, locale_error_zenity, steam_app_factory, caplog):
+            self, locale_error_zenity, steam_app_factory, steam_dir, caplog):
         """
         Try choosing a game with an environment that can't handle non-ASCII
         characters
@@ -128,7 +161,9 @@ class TestSelectApp:
         # Fake user selecting 'Fäke game 2'. The non-ASCII character 'ä'
         # is stripped since Zenity wouldn't be able to display the character.
         locale_error_zenity.mock_stdout = "Fke game 2: 20"
-        steam_app = select_steam_app_with_gui(steam_apps=steam_apps)
+        steam_app = select_steam_app_with_gui(
+            steam_apps=steam_apps, steam_path=steam_dir
+        )
 
         assert steam_app == steam_apps[1]
         assert (
@@ -138,7 +173,8 @@ class TestSelectApp:
 
     @pytest.mark.parametrize("gui_cmd", ["yad", "zenity"])
     def test_select_game_gui_provider_env(
-            self, gui_provider, steam_app_factory, monkeypatch, gui_cmd):
+            self, gui_provider, steam_app_factory, monkeypatch, gui_cmd,
+            steam_dir):
         """
         Test that the correct GUI provider is selected based on the
         `PROTONTRICKS_GUI` environment variable
@@ -151,7 +187,9 @@ class TestSelectApp:
         ]
 
         gui_provider.mock_stdout = "Fake game 2: 20"
-        select_steam_app_with_gui(steam_apps=steam_apps)
+        select_steam_app_with_gui(
+            steam_apps=steam_apps, steam_path=steam_dir
+        )
 
         # The flags should differ slightly depending on which provider is in
         # use