+++ CHANGELOG.md
@@ -10,9 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  - Add desktop integration for Windows executables, which can now be launched using Protontricks
  - Add `protontricks-desktop-install` to install desktop integration for the local user. This is only necessary if the installation method doesn't do this automatically.
  - Add error dialog for displaying error information when Protontricks has been launched from desktop and no user-visible terminal is available.
+ - Add YAD as GUI provider. YAD is automatically used instead of Zenity when available as it supports additional features.
 
 ### Changed
- - Improve GUI dialog. The prompt to select the Steam app now uses zenity's "list" dialog, which supports scrolling and search.
+ - Improved GUI dialog. The prompt to select the Steam app now uses a list dialog with support for scrolling, search and app icons. App icons are only supported on YAD.
 
 ### Fixed
  - Display proper error messages in certain cases when corrupted VDF files are found